Dev express

+++ src/App/Shared/GoldEx.Shared/DTOs/Products/GetGemStoneResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace GoldEx.Shared.DTOs.Products;
 
-public record GetGemStoneResponse(string Code, string Type, string Color, string? Cut, double Carat, string? Purity);
+public record GetGemStoneResponse(string Code, string Type, string Color, string? Cut, decimal Carat, string? Purity);
